backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
celery==5.6.0
redis==7.1.0
psycopg[binary]
python-dotenv==1.2.1